maps: share `Ai_Creaper_Update`
Thanks to @Sezzary for matching it

Weird issue with the map5_s00 version, one of the funcs it calls has an
extra block of code added to the end, wonder what it means...

Maybe map5_s00 was built at a different time to the other maps, and has
this extra block added/removed at some point?
Or could be each map actually has its own copy of the AI code for each
chara? Guess we'll find out once we get more AI stuff done