Unpack `sShSaveGame.enemyKillCountPacked_25C`
Add `q20_12` typedef and use it on `s_ShSaveGame.gameplayTimer`.

Rename `s_ShSaveGame.field_24A` to
`s_ShSaveGame.clearGameCount_24A` (max is 99)

change `s32 enemyKillCountPacked_25C` to separate values.
Looks like `unknwon_25C_3` is still used, and is not padding.

Side Note:
`meleeKillCount` is a weird one. Logically it should be the same size as `rangedKillCount`. But for some reason in the RAM it can be modified as `u16`.
But it can't be packed as `u16`, because the struct will we bigger with added padding.

Need more research.

Also in the future all these 4 `u8` bit field values should be packed to a struct, because it is used again in save-load screen.