bodyprog: 11 matches
Added A LOT of functions that were matched a long time ago in decomp.me. Some of them needed minor touch ups as they were matched with outdated contexts and in one case I added a bit of context for the argument of a function.

Also added one that is almost matched and some that when inserted generate missmatches.

Matches & credits:
- func_80036B5C
@belek666
@Sezzary

- func_800378D4
@belek666
@emoose
@Sezzary

- func_8003BAC4
@belek666
@Sezzary

- func_8003BBF4
@belek666
@Wezax

- func_8003BD48
@emoose
@karas84
@belek666

- func_8003BF60
@belek666
@Sparagas

- func_8003C048
@belek666

- func_8003C0C0
@belek666

- func_8003C110
@fmil95
@belek666

- func_8003C368
@emoose
@Sezzary

- func_8003D5B4
@fmil95
@Sezzary

Co-Authored-By: belek666 <belek666@users.noreply.github.com>
Co-Authored-By: Sezzary <Sezzary@users.noreply.github.com>
Co-Authored-By: Sparagas <Sparagas@users.noreply.github.com>
Co-Authored-By: fmil95 <fmil95@users.noreply.github.com>
Co-Authored-By: emoose <emoose@users.noreply.github.com>
Co-Authored-By: Wezax <Wezax@users.noreply.github.com>
Co-Authored-By: karas84 <karas84@users.noreply.github.com>