Use user in docker container, install some packages (#220)
Create a user with the id, group id and name of the host. This way
the files produced by this container will be owned by the user.
Otherwise all the files produced by the conainer would be owner
by root which is annoying to work with.

Install gcc (needed by permuter) and progress (needed by objdiff)
Install pycparser (needed by permuter)