cleanup s_MapOverlayHeader (#272)
Fix the incorrectly named ".table200Items_4C" and ".always200_50". In
fact this table can have different size in few maps and the second field
is the element count. Same thing with ".table76Items_54" and
".always150_58". Give them a temp "bloodSplats" name as it seems to be
related to that based on memory observation during gameplay.

Replace long 0x00000000 with just 0 and NULL for func pointers.
 #define sharedData symbol alias. Once those are moved to C we can
 delete them.