Macros for screen fade. (#416)
The purpose of this PR is to make simple macros for the most common
screen fade operations used in the game. All of screen fade logic is
controlled by two variables, one of which is a state machine + color bit
combined. Most of the time we just want to start a fade and check if
it's still going on.