Misc: name and objdiff config fix
Adjusted the name in the symbols of a function that was not being counted, and while most changes here are done for the sake of Windows compilation, I noticed that base object building was enabled and that seemly caused objdiff to compile the file despite of having disabled the option of build on changes. Now switching between objects is almost instantaneous lol.